Build with warnings as errors on MSVC, fix some warnings

+++ src/basilisk/RateLimitedWorker.cpp
@@ -6,7 +6,7 @@
 
 namespace chrono = std::chrono;
 
-constexpr int MICROSECONDS_IN_SECOND = chrono::duration_cast<chrono::microseconds>(chrono::seconds(1)).count();
+constexpr uint64_t MICROSECONDS_IN_SECOND = chrono::duration_cast<chrono::microseconds>(chrono::seconds(1)).count();
 
 RateLimitedWorker::RateLimitedWorker(LockBox<Challenge>* box, const Configuration* config, float limit)
 	: super(box, config)
@@ -30,8 +30,8 @@ bool RateLimitedWorker::do_batch()
 
 	float mh = m_speedometer.million_hashes();
 	float seconds = m_speedometer.seconds();
-	float seconds_to_wait = std::max(mh/m_limit - seconds, 0.f);
-	uint64_t microseconds_to_wait = seconds_to_wait * MICROSECONDS_IN_SECOND;
+	double seconds_to_wait = std::max(mh/m_limit - seconds, 0.f);
+	uint64_t microseconds_to_wait = static_cast<uint64_t>(seconds_to_wait * MICROSECONDS_IN_SECOND);
 
 	std::this_thread::sleep_for(chrono::microseconds(microseconds_to_wait));
 	return true;
